batch: Move matcher workspace into package
The line-matching package contains its mapping type while temporary storage
for the matching algorithm remains at the batch package root.

The algorithm and its storage belong to one responsibility, and separating
their locations leaves contributors following an avoidable package
boundary.

This commit advances the package layout by moving matcher workspace storage
beside the mapping type while preserving the selected import path
temporarily.

The next commit will move the matching algorithm onto that workspace.
